Fix build script for use in git tree

+++ build.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
-builddir=/home/pi/compile/solard
+if [ ! -e solard.c ]
+then
+    echo Refusing to start from non-source-root dir...
+    echo ERROR: No solard.c found in `pwd`!
+    exit 2
+fi
 
-cd $builddir
 if [ -e solard ]
 then
     rm solard
@@ -11,5 +15,4 @@ date
 time gcc -D_FORTIFY_SOURCE=2 -Wall -Wno-unused-result -O3 -o solard solard.c
 ls -l
 
-builddir=
 #EOF
